Upgrade worlds to f64 for larger worlds and less loss at farther distances

+++ assets/shaders/daynight_sky.wgsl
@@ -4,7 +4,7 @@ struct Uniforms {
     fog:       vec4<f32>,
     fog_color: vec4<f32>,
     inv_view_proj: mat4x4<f32>,
-    render_origin: vec4<f32>,
+    render_origin: vec4<i32>,
     water_anim: vec4<u32>,
     sky_color: vec4<f32>,
 };
