+++ mods-src/farming/src/compost.rs
@@ -80,10 +80,10 @@ pub fn predict_interact(content: &Content, block: BlockId) -> Outcome {
     }
 }
 
-fn barrel_top(pos: [i32; 3]) -> [f32; 3] {
+fn barrel_top(pos: [i32; 3]) -> [f64; 3] {
     [
-        pos[0] as f32 + 0.5,
-        pos[1] as f32 + 1.2,
-        pos[2] as f32 + 0.5,
+        pos[0] as f64 + 0.5,
+        pos[1] as f64 + 1.2,
+        pos[2] as f64 + 0.5,
     ]
 }
